Fix Some Errors

+++ hw2/Calvin9c/q2/vec_angle.cpp
@@ -6,16 +6,16 @@
 using namespace std;
 namespace py = pybind11;
 
-float vec_angle(float u_x, float u_y, float v_x, float v_y){
-    float u_length = sqrt( u_x*u_x + v_y*v_y );
-    float v_length = sqrt( v_x*v_x + v_y*v_y );
+double vec_angle(double u_x, double u_y, double v_x, double v_y){
+    double u_length = sqrt( u_x*u_x + v_y*v_y );
+    double v_length = sqrt( v_x*v_x + v_y*v_y );
 
-    if ( u_length==0 or v_length==0 ) {
+    if ( u_length==0.0 or v_length==0.0 ) {
         return NAN;
     }
-    float dot_product = u_x*v_x + u_y*v_y;
+    double dot_product = u_x*v_x + u_y*v_y;
 
-    float theta = acos(dot_product/(u_length*v_length));
+    double theta = acos(dot_product/(u_length*v_length));
 
     return theta;
 }
@@ -24,7 +24,7 @@ PYBIND11_MODULE(_vector, m) {
     m.doc() = "Convert c++ program to python one.";
     m.def(
         "vec_angle",
-        &vec_angle,
+        &vec_angle, 
         "return the angle of two 2D-vector."
     );
 }
